requirements.txt
asgiref==3.4.1
boto3==1.18.21
botocore==1.21.21
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
colorama==0.4.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.45.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-recaptcha==2.0.6
django-storages==1.11.1
gunicorn==20.1.0
idna==3.2
jmespath==0.10.0
oauthlib==3.1.1
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
pyotp==2.6.0
python-dateutil==2.8.2
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
